refactor(workers): Move file list storage to common module
Move the `OrtServerFileListStorage` from `:workers:scanner` to
`:workers:common` to prepare for using it in other workers.
